src/requirements.txt
aiogram==2.9.2
Pillow==7.2.0
